Module jet_fixed_term::accounts

An Anchor generated module, providing a set of structs mirroring the structs deriving Accounts, where each field is a Pubkey. This is useful for specifying accounts for a client.

Structs

AuthorizeCrank

Generated client accounts for AuthorizeCrank.

CancelOrder

Generated client accounts for CancelOrder.

ConsumeEvents

Generated client accounts for ConsumeEvents.

ExchangeTokens

Generated client accounts for ExchangeTokens.

InitializeMarginUser

Generated client accounts for InitializeMarginUser.

InitializeMarket

Generated client accounts for InitializeMarket.

InitializeOrderbook

Generated client accounts for InitializeOrderbook.

LendOrder

Generated client accounts for LendOrder.

MarginBorrowOrder

Generated client accounts for MarginBorrowOrder.

MarginLendOrder

Generated client accounts for MarginLendOrder.

MarginRedeemTicket

Generated client accounts for MarginRedeemTicket.

MarginSellTicketsOrder

Generated client accounts for MarginSellTicketsOrder.

ModifyMarket

Generated client accounts for ModifyMarket.

OrderbookMut

Generated client accounts for OrderbookMut.

PauseOrderMatching

Generated client accounts for PauseOrderMatching.

PopAdapterEvents

Generated client accounts for PopAdapterEvents.

RedeemTicket

Generated client accounts for RedeemTicket.

RefreshPosition

Generated client accounts for RefreshPosition.

RegisterAdapter

Generated client accounts for RegisterAdapter.

Repay

Generated client accounts for Repay.

ResumeOrderMatching

Generated client accounts for ResumeOrderMatching.

RevokeCrank

Generated client accounts for RevokeCrank.

SellTicketsOrder

Generated client accounts for SellTicketsOrder.

Settle

Generated client accounts for Settle.

StakeTickets

Generated client accounts for StakeTickets.

TransferTicketOwnership

Generated client accounts for TransferTicketOwnership.