Module jet_fixed_term::jet_fixed_term

Functions

authorize_crank

authorize an address to run orderbook consume_event instructions

cancel_order

Cancels an order on the book

consume_events

Crank specific instruction, processes the event queue

exchange_tokens

Exchange underlying token for fixed term tickets WARNING: tickets must be staked for redeption of underlying

initialize_margin_user

Create a new borrower account

initialize_market

Initializes a Market for a fixed term market

initialize_orderbook

Initializes a new orderbook

lend_order

Place a Lend order to the book by depositing tokens

margin_borrow_order

Place a borrow order by leveraging margin account value

margin_lend_order

Place a Lend order to the book by depositing tokens

margin_redeem_ticket

Redeem a staked ticket

margin_sell_tickets_order

Sell tickets that are already owned

modify_market

Modify a Market account Authority use only

pause_order_matching

Pause matching of orders placed in the orderbook

pop_adapter_events

Pop the given number of events off the adapter queue Event logic is left to the outside program

redeem_ticket

Redeems staked tickets for their underlying value

refresh_position

Refresh the associated margin account claims for a given MarginUser account

register_adapter

Register a new EventAdapter for syncing to the orderbook events

repay

Repay debt on an TermLoan

resume_order_matching

Resume matching of orders placed in the orderbook NOTE: This instruction may have to be run several times to clear the existing matches. Check the orderbook_market_state.pause_matching variable to determine success

revoke_crank

unauthorize an address to run orderbook consume_event instructions

sell_tickets_order

Place an order to the book to sell tickets, which will burn them

settle

Settle payments to a margin account

stake_tickets

Stakes tickets for later redemption

tranfer_ticket_ownership

Transfer staked tickets to a new owner